Move `FeedImageViewModel` to a separate file
