+++ EssentialFeedII/Feed Presentation/FeedImageViewModel.swift
@@ -1,0 +1,21 @@
+//
+//  FeedImageViewModel.swift
+//  EssentialFeedII
+//
+//  Created by Julius on 19/09/2023.
+//
+
+import Foundation
+
+public struct FeedImageViewModel<Image> {
+    public let description: String?
+    public let location: String?
+    public let image: Image?
+    public let isLoading: Bool
+    public let shouldRetry: Bool
+    
+    public var hasLocation: Bool {
+        return location != nil
+    }
+    
+}
